objId is size_t
git-svn-id: file:///home/roberto/Git/svn@7145 1a721e8d-d38c-4e91-ab98-6a0be4d754af

Former-commit-id: 8d6be8d0105d3b76aefc617600497ea561967424

+++ trunk/xmipp/libraries/reconstruction/denoise.cpp
@@ -137,7 +137,7 @@ void ProgDenoise::readParams()
     produceSideInfo();
 }
 
-void ProgDenoise::processImage(const FileName &fnImg, const FileName &fnImgOut, long int objId)
+void ProgDenoise::processImage(const FileName &fnImg, const FileName &fnImgOut, size_t objId)
 {
 	Image<double> img;
     img.readApplyGeo(fnImg,mdIn,objId);
